Move CreatePage logic to view model
