+++ backlog/Views/CreatePage.xaml
@@ -38,7 +38,7 @@
                        Margin="10,0,0,20" TextWrapping="Wrap" />
                 <TextBlock Text="Please follow the hints given in the textbox when entering the name. Do NOT use special characters."
                        Foreground="Yellow" FontStyle="Italic" Margin="10,0,0,20" TextWrapping="Wrap" />
-                <ComboBox Header="Type" PlaceholderText="Choose type" Margin="10,0,0,20" SelectedValue="{x:Bind SelectedType, Mode=TwoWay}"> 
+                <ComboBox Header="Type" PlaceholderText="Choose type" Margin="10,0,0,20" SelectedValue="{x:Bind ViewModel.SelectedType, Mode=TwoWay}"> 
                     <x:String>Film</x:String>
                     <x:String>Album</x:String>
                     <x:String>TV</x:String>
@@ -46,30 +46,30 @@
                     <x:String>Book</x:String>
                 </ComboBox>
                 <TextBlock Text="Name" Margin="10,0,0,20" />
-                <TextBox PlaceholderText="{x:Bind PlaceholderText, Mode=OneWay}" Margin="10,0,0,20" KeyDown="NameInput_KeyDown"
-                         Text="{x:Bind NameInput, Mode=TwoWay}"/>
+                <TextBox PlaceholderText="{x:Bind ViewModel.PlaceholderText, Mode=OneWay}" Margin="10,0,0,20" KeyDown="NameInput_KeyDown"
+                         Text="{x:Bind ViewModel.NameInput, Mode=TwoWay}"/>
                 <TextBlock Text="When do you want to check this out? (optional)" Margin="10,0,0,20" />
-                <CalendarDatePicker Margin="10,0,0,20" Date="{x:Bind DateInput, Mode=TwoWay}" MinDate="{x:Bind today}" />
+                <CalendarDatePicker Margin="10,0,0,20" Date="{x:Bind ViewModel.DateInput, Mode=TwoWay}" MinDate="{x:Bind ViewModel.Today}" />
                 <ToggleSwitch Header="Notify me on this date" Margin="10,0,0,20"
-                              OffContent="No" OnContent="Yes" IsEnabled="{x:Bind EnableNotificationToggle, Mode=OneWay}" IsOn="{x:Bind ShowNotificationOptions, Mode=TwoWay}" />
+                              OffContent="No" OnContent="Yes" IsEnabled="{x:Bind ViewModel.EnableNotificationToggle, Mode=OneWay}" IsOn="{x:Bind ViewModel.ShowNotificationOptions, Mode=TwoWay}" />
                 <TimePicker Header="When do you want to be notified?" 
-                            Margin="10,0,0,20" Visibility="{x:Bind ShowNotificationOptions, Mode=OneWay}" Time="{x:Bind NotifTime, Mode=TwoWay}" />
+                            Margin="10,0,0,20" Visibility="{x:Bind ViewModel.ShowNotificationOptions, Mode=OneWay}" Time="{x:Bind ViewModel.NotifTime, Mode=TwoWay}" />
             </StackPanel>
         </ScrollViewer>
 
-        <muxc:ProgressBar x:Name="progBar" IsIndeterminate="True" Visibility="{x:Bind IsBusy, Mode=OneWay}" />
+        <muxc:ProgressBar x:Name="progBar" IsIndeterminate="True" Visibility="{x:Bind ViewModel.IsBusy, Mode=OneWay}" />
 
-        <ContentDialog x:Name="ResultsDialog" CloseButtonText="Cancel" PrimaryButtonText="Create" PrimaryButtonCommand="{x:Bind CreateBacklog}"
-                       IsPrimaryButtonEnabled="{x:Bind CreateButtonEnabled, Mode=OneWay}">
+        <ContentDialog x:Name="resultsDialog" CloseButtonText="Cancel" PrimaryButtonText="Create" PrimaryButtonCommand="{x:Bind ViewModel.CreateBacklog}"
+                       IsPrimaryButtonEnabled="{x:Bind ViewModel.CreateButtonEnabled, Mode=OneWay}">
             <Grid>
                 <Grid.RowDefinitions>
                     <RowDefinition Height="Auto" />
                     <RowDefinition Height="*" />
                 </Grid.RowDefinitions>
-                
-                <TextBlock Text="{x:Bind SearchResultTitle}" TextWrapping="Wrap" Margin="0,10,0,30" />
-                <ListView SelectedItem="{x:Bind SelectedSearchResult, Mode=TwoWay}" ScrollViewer.VerticalScrollBarVisibility="Auto" Grid.Row="1"
-                          ItemsSource="{x:Bind SearchResults, Mode=OneWay}">
+
+                <TextBlock Text="{x:Bind ViewModel.SearchResultTitle}" TextWrapping="Wrap" Margin="0,10,0,30" />
+                <ListView SelectedItem="{x:Bind ViewModel.SelectedSearchResult, Mode=TwoWay}" ScrollViewer.VerticalScrollBarVisibility="Auto" Grid.Row="1"
+                          ItemsSource="{x:Bind ViewModel.SearchResults, Mode=OneWay}">
                     <ListView.ItemTemplate>
                         <DataTemplate x:DataType="data:SearchResult">
                             <Grid Margin="20">
@@ -92,8 +92,8 @@
     </Grid>
     <Page.TopAppBar>
         <CommandBar x:Name="TopBar" Background="Transparent" DefaultLabelPosition="Right" >
-            <AppBarButton Icon="Find" Label="Search" Command="{x:Bind SearchBacklog}" />
-            <AppBarButton Icon="Cancel" Label="Cancel" Command="{x:Bind Cancel}" />
+            <AppBarButton Icon="Find" Label="Search" Command="{x:Bind ViewModel.SearchBacklog}" />
+            <AppBarButton Icon="Cancel" Label="Cancel" Command="{x:Bind ViewModel.Cancel}" />
             <CommandBar.SecondaryCommands>
                 <AppBarButton x:Name="TopSettingsButton" Icon="Setting" Label="Settings + more" Click="SettingsButton_Click" />
             </CommandBar.SecondaryCommands>
@@ -101,8 +101,8 @@
     </Page.TopAppBar>
     <Page.BottomAppBar>
         <CommandBar x:Name="BottomBar" Visibility="Collapsed" >
-            <AppBarButton Icon="Find" Label="Search" Command="{x:Bind SearchBacklog}" />
-            <AppBarButton Icon="Cancel" Label="Cancel" Command="{x:Bind Cancel}" />
+            <AppBarButton Icon="Find" Label="Search" Command="{x:Bind ViewModel.SearchBacklog}" />
+            <AppBarButton Icon="Cancel" Label="Cancel" Command="{x:Bind ViewModel.Cancel}" />
             <CommandBar.SecondaryCommands>
                 <AppBarButton x:Name="BottomSettingsButton" Icon="Setting" Label="Settings + more" Click="SettingsButton_Click" />
             </CommandBar.SecondaryCommands>
